tests/wpt/web-platform-tests/dom/nodes/CharacterData-deleteData.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>CharacterData.deleteData</title>
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-deletedata">
<link rel=help href="https://dom.spec.whatwg.org/#dom-characterdata-data">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<script>
function testNode(create, type) {
  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    assert_throws("INDEX_SIZE_ERR", function() { node.deleteData(5, 10) })
    assert_throws("INDEX_SIZE_ERR", function() { node.deleteData(5, 0) })
  }, type + ".deleteData() out of bounds")

  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.deleteData(2, 10)
    assert_equals(node.data, "te")
  }, type + ".deleteData() at the end")

  test(function() {
    var node = create()
    assert_equals(node.data, "test")

    node.deleteData(1, 1)
    assert_equals(node.data, "tst")
  }, type + ".deleteData() in the middle")
}

testNode(function() { return document.createTextNode("test") }, "Text")
testNode(function() { return document.createComment("test") }, "Comment")
</script>
